requirements.txt
Flask==2.0.1
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.2.0
gunicorn==20.1.0 